+++ src/main/java/com/hust/mapping/IMediaMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.hust.dao.mapper.IMediaMapper">
-  <resultMap id="BaseResultMap" type="com.hust.model.po.IMedia">
+  <resultMap id="BaseResultMap" type="com.hust.model.IMedia">
     <id column="id" jdbcType="INTEGER" property="id" />
     <result column="name" jdbcType="VARCHAR" property="name" />
     <result column="level" jdbcType="VARCHAR" property="level" />
@@ -67,7 +67,7 @@
   <sql id="Base_Column_List">
     id, name, level
   </sql>
-  <select id="selectByExample" parameterType="com.hust.model.vo.IMediaExample" resultMap="BaseResultMap">
+  <select id="selectByExample" parameterType="com.hust.model.IMediaExample" resultMap="BaseResultMap">
     select
     <if test="distinct">
       distinct
@@ -92,19 +92,19 @@
     delete from media
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByExample" parameterType="com.hust.model.vo.IMediaExample">
+  <delete id="deleteByExample" parameterType="com.hust.model.IMediaExample">
     delete from media
     <if test="_parameter != null">
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="com.hust.model.po.IMedia">
+  <insert id="insert" parameterType="com.hust.model.IMedia">
     insert into media (id, name, level
       )
     values (#{id,jdbcType=INTEGER}, #{name,jdbcType=VARCHAR}, #{level,jdbcType=VARCHAR}
       )
   </insert>
-  <insert id="insertSelective" parameterType="com.hust.model.po.IMedia">
+  <insert id="insertSelective" parameterType="com.hust.model.IMedia">
     insert into media
     <trim prefix="(" suffix=")" suffixOverrides=",">
       <if test="id != null">
@@ -129,7 +129,7 @@
       </if>
     </trim>
   </insert>
-  <select id="countByExample" parameterType="com.hust.model.vo.IMediaExample" resultType="java.lang.Long">
+  <select id="countByExample" parameterType="com.hust.model.IMediaExample" resultType="java.lang.Long">
     select count(*) from media
     <if test="_parameter != null">
       <include refid="Example_Where_Clause" />
@@ -161,7 +161,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="com.hust.model.po.IMedia">
+  <update id="updateByPrimaryKeySelective" parameterType="com.hust.model.IMedia">
     update media
     <set>
       <if test="name != null">
@@ -173,7 +173,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="com.hust.model.po.IMedia">
+  <update id="updateByPrimaryKey" parameterType="com.hust.model.IMedia">
     update media
     set name = #{name,jdbcType=VARCHAR},
       level = #{level,jdbcType=VARCHAR}
