+++ src/main/java/com/hust/mapping/InfoTypeMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.hust.dao.mapper.InfoTypeMapper">
-  <resultMap id="BaseResultMap" type="com.hust.model.po.InfoType">
+  <resultMap id="BaseResultMap" type="com.hust.model.InfoType">
     <id column="id" jdbcType="INTEGER" property="id" />
     <result column="name" jdbcType="VARCHAR" property="name" />
     <result column="weight" jdbcType="INTEGER" property="weight" />
@@ -67,7 +67,7 @@
   <sql id="Base_Column_List">
     id, name, weight
   </sql>
-  <select id="selectByExample" parameterType="com.hust.model.vo.InfoTypeExample" resultMap="BaseResultMap">
+  <select id="selectByExample" parameterType="com.hust.model.InfoTypeExample" resultMap="BaseResultMap">
     select
     <if test="distinct">
       distinct
@@ -92,19 +92,19 @@
     delete from info_type
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByExample" parameterType="com.hust.model.vo.InfoTypeExample">
+  <delete id="deleteByExample" parameterType="com.hust.model.InfoTypeExample">
     delete from info_type
     <if test="_parameter != null">
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="com.hust.model.po.InfoType">
+  <insert id="insert" parameterType="com.hust.model.InfoType">
     insert into info_type (id, name, weight
       )
     values (#{id,jdbcType=INTEGER}, #{name,jdbcType=VARCHAR}, #{weight,jdbcType=INTEGER}
       )
   </insert>
-  <insert id="insertSelective" parameterType="com.hust.model.po.InfoType">
+  <insert id="insertSelective" parameterType="com.hust.model.InfoType">
     insert into info_type
     <trim prefix="(" suffix=")" suffixOverrides=",">
       <if test="id != null">
@@ -129,7 +129,7 @@
       </if>
     </trim>
   </insert>
-  <select id="countByExample" parameterType="com.hust.model.vo.InfoTypeExample" resultType="java.lang.Long">
+  <select id="countByExample" parameterType="com.hust.model.InfoTypeExample" resultType="java.lang.Long">
     select count(*) from info_type
     <if test="_parameter != null">
       <include refid="Example_Where_Clause" />
@@ -161,7 +161,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="com.hust.model.po.InfoType">
+  <update id="updateByPrimaryKeySelective" parameterType="com.hust.model.InfoType">
     update info_type
     <set>
       <if test="name != null">
@@ -173,7 +173,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="com.hust.model.po.InfoType">
+  <update id="updateByPrimaryKey" parameterType="com.hust.model.InfoType">
     update info_type
     set name = #{name,jdbcType=VARCHAR},
       weight = #{weight,jdbcType=INTEGER}
