+++ src/main/java/com/hust/mapping/LMediaMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.hust.dao.mapper.LMediaMapper">
-  <resultMap id="BaseResultMap" type="com.hust.model.po.LMedia">
+  <resultMap id="BaseResultMap" type="com.hust.model.LMedia">
     <id column="id" jdbcType="INTEGER" property="id" />
     <result column="name" jdbcType="VARCHAR" property="name" />
     <result column="weight" jdbcType="INTEGER" property="weight" />
@@ -67,7 +67,7 @@
   <sql id="Base_Column_List">
     id, name, weight
   </sql>
-  <select id="selectByExample" parameterType="com.hust.model.vo.LMediaExample" resultMap="BaseResultMap">
+  <select id="selectByExample" parameterType="com.hust.model.LMediaExample" resultMap="BaseResultMap">
     select
     <if test="distinct">
       distinct
@@ -92,19 +92,19 @@
     delete from level
     where id = #{id,jdbcType=INTEGER}
   </delete>
-  <delete id="deleteByExample" parameterType="com.hust.model.vo.LMediaExample">
+  <delete id="deleteByExample" parameterType="com.hust.model.LMediaExample">
     delete from level
     <if test="_parameter != null">
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" parameterType="com.hust.model.po.LMedia">
+  <insert id="insert" parameterType="com.hust.model.LMedia">
     insert into level (id, name, weight
       )
     values (#{id,jdbcType=INTEGER}, #{name,jdbcType=VARCHAR}, #{weight,jdbcType=INTEGER}
       )
   </insert>
-  <insert id="insertSelective" parameterType="com.hust.model.po.LMedia">
+  <insert id="insertSelective" parameterType="com.hust.model.LMedia">
     insert into level
     <trim prefix="(" suffix=")" suffixOverrides=",">
       <if test="id != null">
@@ -129,7 +129,7 @@
       </if>
     </trim>
   </insert>
-  <select id="countByExample" parameterType="com.hust.model.vo.LMediaExample" resultType="java.lang.Long">
+  <select id="countByExample" parameterType="com.hust.model.LMediaExample" resultType="java.lang.Long">
     select count(*) from level
     <if test="_parameter != null">
       <include refid="Example_Where_Clause" />
@@ -161,7 +161,7 @@
       <include refid="Update_By_Example_Where_Clause" />
     </if>
   </update>
-  <update id="updateByPrimaryKeySelective" parameterType="com.hust.model.po.LMedia">
+  <update id="updateByPrimaryKeySelective" parameterType="com.hust.model.LMedia">
     update level
     <set>
       <if test="name != null">
@@ -173,7 +173,7 @@
     </set>
     where id = #{id,jdbcType=INTEGER}
   </update>
-  <update id="updateByPrimaryKey" parameterType="com.hust.model.po.LMedia">
+  <update id="updateByPrimaryKey" parameterType="com.hust.model.LMedia">
     update level
     set name = #{name,jdbcType=VARCHAR},
       weight = #{weight,jdbcType=INTEGER}
