+++ src/main/java/com/hust/model/IMedia.java
@@ -1,0 +1,33 @@
+package com.hust.model;
+
+public class IMedia {
+    private Integer id;
+
+    private String name;
+
+    private String level;
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name == null ? null : name.trim();
+    }
+
+    public String getLevel() {
+        return level;
+    }
+
+    public void setLevel(String level) {
+        this.level = level == null ? null : level.trim();
+    }
+}
