+++ src/main/java/com/hust/model/InfoType.java
@@ -1,0 +1,33 @@
+package com.hust.model;
+
+public class InfoType {
+    private Integer id;
+
+    private String name;
+
+    private Integer weight;
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name == null ? null : name.trim();
+    }
+
+    public Integer getWeight() {
+        return weight;
+    }
+
+    public void setWeight(Integer weight) {
+        this.weight = weight;
+    }
+}
